Move Swahili digit cleaners to a language-independent DigitCleaner class
