+++ core_data_modules/cleaners/__init__.py
@@ -1,4 +1,5 @@
 from .character_cleaner import CharacterCleaner
 from .codes import Codes
+from .digit_cleaner import DigitCleaner
 from .phone_cleaner import PhoneCleaner
 from .regex_utils import RegexUtils
